refactor: split ingestion lifecycle dtos
